+++ models/index.js
@@ -1,0 +1,7 @@
+const inventory = require("./inventory");
+const invoice = require("./invoice");
+const location = require("./location");
+const payments = require("./payments");
+const personnel = require("./personnel");
+
+module.exports = [inventory, invoice, location, payments, personnel];
