+++ models/inventory/index.js
@@ -1,0 +1,5 @@
+const batch = require("./batch");
+const inventory = require("./inventory");
+const product = require("./product");
+
+module.exports = [batch, inventory, product];
