+++ models/invoice/index.js
@@ -1,0 +1,4 @@
+const invoice = require("./invoice");
+const invoiceItem = require("./invoiceItem");
+
+module.exports = [invoice, invoiceItem];
