+++ models/location/index.js
@@ -1,0 +1,9 @@
+const location = require("./location");
+const building = require("./building");
+const floor = require("./floor");
+const room = require("./room");
+const section = require("./section");
+const subSection = require("./subSection");
+const shelf = require("./shelf");
+
+module.exports = [location, building, floor, room, section, subSection, shelf];
