+++ models/payments/index.js
@@ -1,0 +1,6 @@
+const payment = require("./payment");
+const transaction = require("./transaction");
+const transactionType = require("./transactionType");
+const financialTransaction = require("./financialTransaction");
+
+module.exports = [payment, transaction, transactionType, financialTransaction];
