+++ models/personnel/index.js
@@ -1,0 +1,4 @@
+const staff = require("./staff");
+const hourlyRate = require("./hourlyRate");
+
+module.exports = [staff, hourlyRate];
